Gestion de Droit

+++ backend/src/main/java/com/example/backendagile/mapper/DroitMapper.java
@@ -54,7 +54,7 @@ public class DroitMapper {
         Optional<Evaluation> eval = Optional.ofNullable(evaluationService.getEvaluationByID(droitDTO.getIdEvaluation()));
 
 
-        droit.setId(new DroitId(droitDTO.getIdEvaluation(), droitDTO.getIdEnseignant().intValue()));
+        droit.setId(new DroitId(droitDTO.getIdEvaluation(), droitDTO.getIdEnseignant()));
         droit.setIdEvaluation(eval.orElse(null));
         droit.setNoEnseignant(ens.orElse(null));
         droit.setConsultation(Character.toUpperCase(droitDTO.getConsultation()));
